pyproject.toml
[project]
name = "gdocs-to-web"
version = "0.1.0"
package-mode = false
description = "A simple script to convert Google Docs Markdown export to valid HTML."
authors = [
    {name = "Markus Pettersson",email = "markus.pettersson@edu.huddinge.se"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "markdown (>=3.9,<4.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "python-slugify (>=8.0.4,<9.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
